# xxh32: ARM assembler implementation