Here's an ARM assembler version of xxhash32. The patch is rather big, as I had to shuffle some code around. golang.org/x/sys is now a runtime dependency for any clients, because we need to see the CPU features.

Benchmark results on Raspberry Pi 4B compared to old code (with GOARM=7):

```
name                old time/op    new time/op    delta
pkg:github.com/pierrec/lz4/v4/internal/xxh32 goos:linux goarch:arm
_XXH32-4               811ns ± 1%     291ns ± 1%   -64.14%  (p=0.008 n=5+5)
_XXH32_Checksum-4      736ns ± 1%     176ns ± 0%   -76.09%  (p=0.016 n=5+4)

name                old speed      new speed      delta
pkg:github.com/pierrec/lz4/v4 goos:linux goarch:arm
CompressRandom-4     131MB/s ± 0%   131MB/s ± 1%      ~     (p=0.690 n=5+5)
UncompressPg1661-4   138MB/s ± 1%   161MB/s ± 0%   +16.71%  (p=0.016 n=5+4)
UncompressDigits-4   272MB/s ± 2%   403MB/s ± 2%   +48.23%  (p=0.008 n=5+5)
UncompressTwain-4    142MB/s ± 2%   167MB/s ± 3%   +17.24%  (p=0.008 n=5+5)
UncompressRand-4     478MB/s ± 1%  1157MB/s ± 2%  +142.11%  (p=0.008 n=5+5)
```

Only ARMv7 is supported. The pure Go version is used as a fallback for ARMv{5,6} (and ARMv7 without NEON, due to how we check for ARMv7). Most of the speedup is in replacing the 16 byte loads + ORRs by a single MOVM, so the inner loop goes from 42 instructions to 15. Without this, ChecksumZero is only ~20% faster in assembly.

Performance on amd64 is unchanged, despite the additional function call due to code restructuring.